backend/llm-service/pyproject.toml
[tool.poetry]
name = "llm-service"
version = "0.1.0"
description = "LLM service for ANCAP project"
authors = ["Santiago Morón Santos <111464779+Santiago-Moron@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "llm_service" }]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.09,<3.14"
fastapi = "^0.115.0"
langsmith = "^0.3.44"
langchain-openai = "^0.3.19"
langgraph = "^0.4.8"
dotenv = "^0.9.9"
uvicorn = "^0.34.3"
psycopg2-binary = "^2.9.10"
langchain="~0.3.25"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
